ZeroShotEvaluation/requirements.txt
torch
transformers
accelerate
bitsandbytes
Pillow
pandas
tqdm
requests
pycocoevalcap